+++ src/number_theory/gcd.rs
@@ -1,5 +1,5 @@
 // Compute the GCD of m & n using Euclid's Algorithm
-pub fn gcd(m: i32, n: i32) -> i32 {
+pub fn gcd(m: i64, n: i64) -> i64 {
     if n == 0 {
         return m.abs();
     }
@@ -7,7 +7,7 @@ pub fn gcd(m: i32, n: i32) -> i32 {
 }
 
 // Compute the GCD of a list
-pub fn gcd_arr(v: &[i32]) -> i32 {
+pub fn gcd_arr(v: &[i64]) -> i64 {
     let mut gcd_r = v[0];
     for num in v {
         gcd_r = gcd(gcd_r, *num);
